# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  /** Local phone number without country code */
  phoneNumber: string;
  /** E.164 country dial code (e.g. "+1", "+44", "+91") */
  countryCode: string;
  /**
     * ISO 3166-1 alpha-2 country code (e.g. "US", "GB", "IN")
     * @nullable
     */
  countryIso?: string | null;
  /** Full phone in E.164 format (e.g. "+14155552671") */
  fullPhone?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phoneNumber: string;
  /** @minLength 1 */
  countryCode: string;
  /** @minLength 2 */
  countryIso: string;
}

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phoneNumber?: string;
  countryCode?: string;
  countryIso?: string;
}

export type ConsentType = typeof ConsentType[keyof typeof ConsentType];


export const ConsentType = {
  location: 'location',
  notification: 'notification',
  messaging: 'messaging',
} as const;

export type ConsentStatus = typeof ConsentStatus[keyof typeof ConsentStatus];


export const ConsentStatus = {
  granted: 'granted',
  denied: 'denied',
  revoked: 'revoked',
} as const;

export interface Consent {
  id: number;
  userId: number;
  type: ConsentType;
  status: ConsentStatus;
  /**
     * Human-readable reason for this consent
     * @nullable
     */
  purpose?: string | null;
  /** @nullable */
  grantedAt?: string | null;
  /** @nullable */
  revokedAt?: string | null;
  createdAt: string;
}

export type ConsentInputType = typeof ConsentInputType[keyof typeof ConsentInputType];


export const ConsentInputType = {
  location: 'location',
  notification: 'notification',
  messaging: 'messaging',
} as const;

export type ConsentInputStatus = typeof ConsentInputStatus[keyof typeof ConsentInputStatus];


export const ConsentInputStatus = {
  granted: 'granted',
  denied: 'denied',
  revoked: 'revoked',
} as const;

export interface ConsentInput {
  userId: number;
  type: ConsentInputType;
  status: ConsentInputStatus;
  purpose?: string;
}

export type ConsentUpdateStatus = typeof ConsentUpdateStatus[keyof typeof ConsentUpdateStatus];


export const ConsentUpdateStatus = {
  granted: 'granted',
  denied: 'denied',
  revoked: 'revoked',
} as const;

export interface ConsentUpdate {
  status?: ConsentUpdateStatus;
  purpose?: string;
}

export interface ConsentTypeStat {
  granted: number;
  denied: number;
  revoked: number;
  total: number;
}

export interface ConsentSummary {
  location: ConsentTypeStat;
  notification: ConsentTypeStat;
  messaging: ConsentTypeStat;
  totals: ConsentTypeStat;
}

export type InviteStatus = typeof InviteStatus[keyof typeof InviteStatus];


export const InviteStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

/**
 * Which permission this invite is requesting
 * @nullable
 */
export type InviteConsentType = typeof InviteConsentType[keyof typeof InviteConsentType] | null;


export const InviteConsentType = {
  location: 'location',
  notification: 'notification',
  messaging: 'messaging',
} as const;

export interface Invite {
  id: number;
  fromUserId: number;
  toPhone: string;
  /** @nullable */
  toName?: string | null;
  message: string;
  status: InviteStatus;
  /** Generated wa.me deep link */
  whatsappLink: string;
  /**
     * Which permission this invite is requesting
     * @nullable
     */
  consentType?: InviteConsentType;
  sentAt: string;
}

export type InviteInputConsentType = typeof InviteInputConsentType[keyof typeof InviteInputConsentType];


export const InviteInputConsentType = {
  location: 'location',
  notification: 'notification',
  messaging: 'messaging',
} as const;

export interface InviteInput {
  fromUserId: number;
  /** @minLength 1 */
  toPhone: string;
  toName?: string;
  /** @minLength 1 */
  message: string;
  consentType?: InviteInputConsentType;
}

export type InviteUpdateStatus = typeof InviteUpdateStatus[keyof typeof InviteUpdateStatus];


export const InviteUpdateStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface InviteUpdate {
  status?: InviteUpdateStatus;
  message?: string;
}

export interface WhatsappLinkRequest {
  /** Phone number in E.164 format (digits only, no +) */
  phoneNumber: string;
  /** Pre-filled message text */
  message: string;
}

export interface WhatsappLink {
  /** Full wa.me deep link URL */
  link: string;
  phoneNumber: string;
  message: string;
}

export type ListConsentsParams = {
userId?: number;
type?: ListConsentsType;
status?: ListConsentsStatus;
};

export type ListConsentsType = typeof ListConsentsType[keyof typeof ListConsentsType];


export const ListConsentsType = {
  location: 'location',
  notification: 'notification',
  messaging: 'messaging',
} as const;

export type ListConsentsStatus = typeof ListConsentsStatus[keyof typeof ListConsentsStatus];


export const ListConsentsStatus = {
  granted: 'granted',
  denied: 'denied',
  revoked: 'revoked',
} as const;

export type ListInvitesParams = {
userId?: number;
status?: ListInvitesStatus;
};

export type ListInvitesStatus = typeof ListInvitesStatus[keyof typeof ListInvitesStatus];


export const ListInvitesStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;