# lib/api-zod/src/generated/types/consentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */
import type { ConsentInputStatus } from './consentInputStatus';
import type { ConsentInputType } from './consentInputType';

export interface ConsentInput {
  userId: number;
  type: ConsentInputType;
  status: ConsentInputStatus;
  purpose?: string;
}