# lib/api-zod/src/generated/types/consentTypeStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentTypeStat {
  granted: number;
  denied: number;
  revoked: number;
  total: number;
}