# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */
import type { InviteConsentType } from './inviteConsentType';
import type { InviteStatus } from './inviteStatus';

export interface Invite {
  id: number;
  fromUserId: number;
  toPhone: string;
  /** @nullable */
  toName?: string | null;
  message: string;
  status: InviteStatus;
  /** Generated wa.me deep link */
  whatsappLink: string;
  /**
     * Which permission this invite is requesting
     * @nullable
     */
  consentType?: InviteConsentType;
  sentAt: Date;
}