# lib/api-zod/src/generated/types/inviteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */
import type { InviteUpdateStatus } from './inviteUpdateStatus';

export interface InviteUpdate {
  status?: InviteUpdateStatus;
  message?: string;
}