# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** Local phone number without country code */
  phoneNumber: string;
  /** E.164 country dial code (e.g. "+1", "+44", "+91") */
  countryCode: string;
  /**
     * ISO 3166-1 alpha-2 country code (e.g. "US", "GB", "IN")
     * @nullable
     */
  countryIso?: string | null;
  /** Full phone in E.164 format (e.g. "+14155552671") */
  fullPhone?: string;
  createdAt: Date;
  updatedAt?: Date;
}