# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhoneLink API — international phone registration, permission consents, and WhatsApp sharing
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phoneNumber?: string;
  countryCode?: string;
  countryIso?: string;
}